Code Quality: Continued work of applying one class per file (#12241)
