+++ src/Files.App/Data/EventArguments/BundlesOpenPathEventArgs.cs
@@ -1,7 +1,7 @@
 // Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-namespace Files.App.Data.EventArguments.Bundles
+namespace Files.App.Data.EventArguments
 {
 	public class BundlesOpenPathEventArgs
 	{
