+++ src/Files.App/Data/EventArguments/PathNavigationEventArgs.cs
@@ -1,0 +1,18 @@
+﻿// Copyright (c) 2023 Files Community
+// Licensed under the MIT License. See the LICENSE.
+
+using Files.App.Views;
+using Microsoft.UI.Xaml.Controls;
+using Windows.ApplicationModel.DataTransfer;
+
+namespace Files.App.Data.EventArguments
+{
+	public class PathNavigationEventArgs
+	{
+		public string ItemPath { get; set; }
+
+		public string ItemName { get; set; }
+
+		public bool IsFile { get; set; }
+	}
+}
