+++ src/Files.App/Data/EventArguments/ToolbarPathItemLoadedEventArgs.cs
@@ -1,0 +1,16 @@
+﻿// Copyright (c) 2023 Files Community
+// Licensed under the MIT License. See the LICENSE.
+
+using Files.App.Views;
+using Microsoft.UI.Xaml.Controls;
+using Windows.ApplicationModel.DataTransfer;
+
+namespace Files.App.Data.EventArguments
+{
+	public class ToolbarPathItemLoadedEventArgs
+	{
+		public MenuFlyout OpenedFlyout { get; set; }
+
+		public PathBoxItem Item { get; set; }
+	}
+}
