+++ src/Files.App/Data/Factories/AppThemeResourcesFactory.cs
@@ -1,108 +1,108 @@
 ﻿// Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-namespace Files.App.Views.Settings.Appearance
+namespace Files.App.Data.Factories
 {
 	public static class AppThemeResourceFactory
 	{
-		public static ObservableCollection<AppThemeResource> AppThemeResources { get; } = new ObservableCollection<AppThemeResource>()
+		public static ObservableCollection<AppThemeResourceItem> AppThemeResources { get; } = new ObservableCollection<AppThemeResourceItem>()
 		{
-			new AppThemeResource
+			new AppThemeResourceItem
 			{
 				BackgroundColor = "#00000000", /* Transparent */
 				Name = "Default"
 			},
-			new AppThemeResource
+			new AppThemeResourceItem
 			{
 				BackgroundColor = "#32FFB900", /* #FFB900 */
 				Name = "Yellow Gold"
 			},
-			new AppThemeResource
+			new AppThemeResourceItem
 			{
 				BackgroundColor = "#32F7630C", /* #F7630C */
 				Name = "Orange Bright"
 			},
-			new AppThemeResource
+			new AppThemeResourceItem
 			{
 				BackgroundColor = "#32D13438", /* #D13438 */
 				Name = "Brick Red"
 			},
-			new AppThemeResource
+			new AppThemeResourceItem
 			{
 				BackgroundColor = "#32FF4343", /* #FF4343 */
 				Name = "Mod Red"
 			},
-			new AppThemeResource
+			new AppThemeResourceItem
 			{
 				BackgroundColor = "#32EA005E", /* #EA005E */
 				Name = "Red"
 			},
-			new AppThemeResource
+			new AppThemeResourceItem
 			{
 				BackgroundColor = "#32EA005E", /* #EA005E */
 				Name = "Rose Bright"
 			},
-			new AppThemeResource
+			new AppThemeResourceItem
 			{
 				BackgroundColor = "#320078D7", /* #0078D7 */
 				Name = "Blue"
 			},
-			new AppThemeResource
+			new AppThemeResourceItem
 			{
 				BackgroundColor = "#328764B8", /* #8764B8 */
 				Name = "Iris Pastel"
 			},
-			new AppThemeResource
+			new AppThemeResourceItem
 			{
 				BackgroundColor = "#32B146C2", /* #B146C2 */
 				Name = "Violet Red Light"
 			},
-			new AppThemeResource
+			new AppThemeResourceItem
 			{
 				BackgroundColor = "#320099BC", /* #0099BC */
 				Name = "Cool Blue Bright"
 			},
-			new AppThemeResource
+			new AppThemeResourceItem
 			{
 				BackgroundColor = "#3200B7C3", /* #00B7C3 */
 				Name = "Seafoam"
 			},
-			new AppThemeResource
+			new AppThemeResourceItem
 			{
 				BackgroundColor = "#3200B294", /* #00B294 */
 				Name = "Mint Light"
 			},
-			new AppThemeResource
+			new AppThemeResourceItem
 			{
 				BackgroundColor = "#327A7574", /* #7A7574 */
 				Name = "Gray"
 			},
-			new AppThemeResource
+			new AppThemeResourceItem
 			{
 				BackgroundColor = "#32107C10", /* #107C10 */
 				Name = "Green"
 			},
-			new AppThemeResource
+			new AppThemeResourceItem
 			{
 				BackgroundColor = "#32767676", /* #767676 */
 				Name = "Overcast"
 			},
-			new AppThemeResource
+			new AppThemeResourceItem
 			{
 				BackgroundColor = "#324C4A48", /* #4C4A48 */
 				Name = "Storm"
 			},
-			new AppThemeResource
+			new AppThemeResourceItem
 			{
 				BackgroundColor = "#3269797E", /* #69797E */
 				Name = "Blue Gray"
 			},
-			new AppThemeResource
+			new AppThemeResourceItem
 			{
 				BackgroundColor = "#324A5459", /* #4A5459 */
 				Name = "Gray Dark"
 			},
-			new AppThemeResource
+			new AppThemeResourceItem
 			{
 				BackgroundColor = "#327E735F", /* #7E735F */
 				Name = "Camouflage"
