+++ src/Files.App/Data/Items/AppThemeResourceItem.cs
@@ -1,11 +1,12 @@
 ﻿// Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-namespace Files.App.Views.Settings.Appearance
+namespace Files.App.Data.Items
 {
-	public class AppThemeResource
+	public class AppThemeResourceItem
 	{
 		public string? Name { get; set; }
+
 		public string? BackgroundColor { get; set; }
 	}
 }
