+++ src/Files.App/Data/Parameters/ColumnParam.cs
@@ -1,9 +1,11 @@
 // Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
+
+using Files.App.Views;
 using Files.App.Views.LayoutModes;
 using Microsoft.UI.Xaml.Controls;
 
-namespace Files.App.Views
+namespace Files.App.Data.Parameters
 {
 	public class ColumnParam : NavigationArguments
 	{
