+++ src/Files.App/ISearchBox.cs
@@ -1,8 +1,6 @@
 // Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-using Files.App.Data.Models;
-using Microsoft.UI.Xaml.Controls;
 using Windows.Foundation;
 
 namespace Files.App
@@ -24,38 +22,5 @@ namespace Files.App
 		void SetSuggestions(IEnumerable<SuggestionModel> suggestions);
 
 		void AddRecentQueries();
-	}
-
-	public class SearchBoxTextChangedEventArgs
-	{
-		public SearchBoxTextChangeReason Reason { get; }
-
-		public SearchBoxTextChangedEventArgs(SearchBoxTextChangeReason reason)
-			=> Reason = reason;
-
-		public SearchBoxTextChangedEventArgs(AutoSuggestionBoxTextChangeReason reason)
-		{
-			Reason = reason switch
-			{
-				AutoSuggestionBoxTextChangeReason.UserInput => SearchBoxTextChangeReason.UserInput,
-				AutoSuggestionBoxTextChangeReason.SuggestionChosen => SearchBoxTextChangeReason.SuggestionChosen,
-				_ => SearchBoxTextChangeReason.ProgrammaticChange
-			};
-		}
-	}
-
-	public class SearchBoxQuerySubmittedEventArgs
-	{
-		public SuggestionModel ChosenSuggestion { get; }
-
-		public SearchBoxQuerySubmittedEventArgs(SuggestionModel chosenSuggestion)
-			=> ChosenSuggestion = chosenSuggestion;
-	}
-
-	public enum SearchBoxTextChangeReason : ushort
-	{
-		UserInput,
-		ProgrammaticChange,
-		SuggestionChosen,
 	}
 }
