+++ src/Files.App/UserControls/IAddressToolbar.cs
@@ -47,46 +47,4 @@ namespace Files.App.UserControls
 
 		public ISearchBox SearchBox { get; }
 	}
-
-	public class ToolbarQuerySubmittedEventArgs
-	{
-		public string QueryText { get; set; } = null;
-	}
-
-	public class PathNavigationEventArgs
-	{
-		public string ItemPath { get; set; }
-
-		public string ItemName { get; set; }
-
-		public bool IsFile { get; set; }
-	}
-
-	public class ToolbarFlyoutOpenedEventArgs
-	{
-		public MenuFlyout OpenedFlyout { get; set; }
-	}
-
-	public class ToolbarPathItemLoadedEventArgs
-	{
-		public MenuFlyout OpenedFlyout { get; set; }
-
-		public PathBoxItem Item { get; set; }
-	}
-
-	public class AddressBarTextEnteredEventArgs
-	{
-		public AutoSuggestBox AddressBarTextField { get; set; }
-	}
-
-	public class PathBoxItemDroppedEventArgs
-	{
-		public DataPackageView Package { get; set; }
-
-		public string Path { get; set; }
-
-		public DataPackageOperation AcceptedOperation { get; set; }
-
-		public AsyncManualResetEvent SignalEvent { get; set; }
-	}
 }
