+++ src/Files.App/ViewModels/Widgets/Bundles/BundlesViewModel.cs
@@ -1,7 +1,6 @@
 // Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-using Files.App.Data.EventArguments.Bundles;
 using Files.App.Dialogs;
 using Files.App.ViewModels.Dialogs;
 using Microsoft.UI.Xaml;
