+++ src/Files.Backend/Enums/SearchBoxTextChangeReason.cs
@@ -1,0 +1,11 @@
+﻿namespace Files.Backend.Enums
+{
+	public enum SearchBoxTextChangeReason : ushort
+	{
+		UserInput,
+
+		ProgrammaticChange,
+
+		SuggestionChosen
+	}
+}
